move radio group css files
